@@ -26,21 +26,30 @@ struct SendFeedbackView: View {
     /// This State variable is updated to true when sending feedback fails.
     @State private var showingAlert = false
     
+    @State private var showLoginView = false
     
+    @State private var finishLoggingIn = false
+    
+    @State private var dismissView = false
     var body: some View {
         VStack {
-            ActivityIndicator(isAnimating: self.$shouldAnimate)
-                .onAppear { self.sendFeedback() }
-            Text("Sending the feedback to Prototyper")
-        }
-            .alert(isPresented: $showingAlert) {
-                Alert(title: Text("Error"),
-                      message: Text("Could not send feedback to server."),
-                      dismissButton: .default(Text("OK")) {
-                          self.showSendFeedbackView = false
-                      })
+            if !(state.userIsLoggedIn || state.continueWithoutLogin) {
+                LoginView()
+            } else {
+                ProgressView {
+                    Text("Sending the feedback to Prototyper")
+                }.onAppear {
+                    self.sendFeedback()
+                }.alert(isPresented: $showingAlert) {
+                    Alert(title: Text("Error"),
+                          message: Text("Could not send feedback to server."),
+                          dismissButton: .default(Text("OK")) {
+                            self.showSendFeedbackView = false
+                          })
+                }
             }
-            .navigationBarTitle("Sending Feedback")
+            
+        }.navigationTitle("Feedback")
     }
     
     
@@ -49,7 +58,7 @@ struct SendFeedbackView: View {
         guard var feedback = feedback else {
             return
         }
-        
+        print("SendFeedback called")
         feedback.creatorName = UserDefaults.standard.string(forKey: UserDefaultKeys.username)
         
         apiHandler.send(
