@@ -11,16 +11,7 @@ import SwiftUI
 // MARK: FeedbackView
 /// This View holds the markup image and the feedback text field for the user to send feedback.
 struct FeedbackView: View {
-    // MARK: ActiveSheet
-    /// The type of active sheet that can be presented by the `FeedbackView`
-    enum ActiveSheet {
-        /// The login sheet is presented
-        case loginSheet
-        /// The markup sheet is presented
-        case markupSheet
-    }
-    
-    
+
     /// The instance of the Observable Object class named Model,  to share state data anywhere it’s needed.
     @EnvironmentObject var state: PrototyperState
     /// <#Description#>
@@ -32,8 +23,6 @@ struct FeedbackView: View {
     @State var showSheet: Bool = false
     /// This State variable is used to tell the View if the screenshot needs to be sent as feedback or not.
     @State var showScreenshot: Bool = true
-    /// The active Sheet decides if the login View or the SendFeedback View needs to be shown when the showSheet variable is updated.
-    @State var activeSheet: ActiveSheet = .loginSheet
     /// This State variable when set to true sends the Feedback to Prototyper.
     @State var showSendFeedbackView: Bool = false
     /// The variable holds the image and the feedback text to be sent.
@@ -59,19 +48,11 @@ struct FeedbackView: View {
             .navigationBarTitle("Give Feedback")
             .navigationBarItems(leading: cancelButton, trailing: sendButton)
             .sheet(isPresented: $showSheet) {
-                if self.activeSheet == .loginSheet {
-                    NavigationView {
-                        LoginView(finishLoggingIn: self.$showSendFeedbackView)
-                    }
-                        .environmentObject(self.state)
-                        .environmentObject(self.apiHandler)
-                } else {
                     NavigationView {
                         EditScreenshotView()
                     }
                         .environmentObject(self.state)
                         .environmentObject(self.apiHandler)
-                }
             }
         }
     }
@@ -103,7 +84,7 @@ struct FeedbackView: View {
     /// Holds the screenshot and the feedback text to be sent to Prototyper
     var currentFeedback: Feedback {
         var creatorName: String?
-        if !apiHandler.isLoggedIn {
+        if state.userIsLoggedIn {
             creatorName = UserDefaults.standard.string(forKey: UserDefaultKeys.username)
         }
         
@@ -136,19 +117,17 @@ struct FeedbackView: View {
     /// Sends the feedback to Prototyper via the SendFeedbackView
     private func send() {
         feedback = currentFeedback
-        if apiHandler.isLoggedIn || state.continueWithoutLogin {
-            state.continueWithoutLogin = false
-            self.showSendFeedbackView = true
-        } else {
-            activeSheet = .loginSheet
-            self.showSheet = true
+        //TODO
+        self.showSendFeedbackView = true
+        NavigationView {
+            SendFeedbackView(showSendFeedbackView: $showSendFeedbackView, feedback: $feedback)
         }
+        
     }
     
     /// Brings up the EditScreenShotView to Markup the screenshot
     private func editImage() {
-        activeSheet = .markupSheet
-        showSheet = true
+            showSheet = true
     }
     
     /// Deletes the screenshot from the View
